Remove casts to f32 for text sizes (#741)
The main change is that text sizes in Masonry::theme are now `f32`. The
rest of the changes are autofixed.

+++ masonry/src/text/text_layout.rs
@@ -607,6 +607,6 @@ impl std::fmt::Debug for TextLayout {
 
 impl Default for TextLayout {
     fn default() -> Self {
-        Self::new(crate::theme::TEXT_SIZE_NORMAL as f32)
+        Self::new(crate::theme::TEXT_SIZE_NORMAL)
     }
 }
